Synchronizing post addition and fixing some bugs.

+++ Backend/controllers/feed.js
@@ -3,8 +3,10 @@ const fs = require("fs");
 
 const { validationResult } = require("express-validator");
 
+const io = require("../socket");
 const Post = require("../models/post");
 const User = require("../models/user");
+const user = require("../models/user");
 
 exports.getPosts = (req, res, next) => {
   const currentPage = req.query.page || 1;
@@ -15,6 +17,7 @@ exports.getPosts = (req, res, next) => {
     .then((count) => {
       totalItems = count;
       return Post.find()
+        .populate("creator")
         .skip((currentPage - 1) * perPage)
         .limit(perPage);
     })
@@ -68,6 +71,10 @@ exports.createPost = (req, res, next) => {
       return user.save();
     })
     .then((result) => {
+      io.getIO().emit("posts", {
+        action: "create",
+        post: { ...post._doc, creator: { _id: req.userId, name: creator.name } },
+      });
       res.status(201).json({
         message: "Post created successfully!",
         post: post,
@@ -176,6 +183,9 @@ exports.deletePost = (req, res, next) => {
         throw error;
       }
       clearImage(post.imageUrl);
+      if (!post.imageUrl) {
+        return;
+      }
       return Post.findByIdAndDelete(postId);
     })
     .then((result) => {
